backend/requirements.txt
# =============================================================================
# DomainRecon - Dépendances Python
# =============================================================================

fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
httpx==0.25.2
tldextract==5.1.1
python-whois==0.9.4
